requirements.txt
# Azure Functions dependencies
azure-functions
azurefunctions-extensions-bindings-blob

# Azure SDK dependencies
azure-cosmos==4.5.1
azure-identity
azure-mgmt-web
azure-mgmt-storage
azure-storage-blob

# AI and NLP dependencies
tiktoken==0.7.0
openai==1.42.0

# GenAI dependencies
pyautogen==0.3.0

aiohttp==3.10.5

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
